pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "anip"
version = "1.0.0"
description = "Automated News Intelligence Pipeline"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}

dependencies = [
    "psycopg2-binary==2.9.9",
    "sqlalchemy==1.4.52",
    "python-dotenv==1.0.0",
    "requests==2.31.0",
    "pydantic==2.5.0",
    "pydantic-settings==2.1.0",
]

[project.optional-dependencies]
airflow = [
    "apache-airflow==2.9.3",
    "newsapi-python==0.2.7",
]

ml = [
    "mlflow==2.16.0",
    "scikit-learn==1.5.0",
    "numpy==1.26.4",
    "scipy==1.13.0",
]

api = [
    "fastapi==0.104.1",
    "uvicorn==0.24.0",
]

spark = [
    "pyspark==4.0.0",
]

dev = [
    "pytest==7.4.3",
    "pytest-cov==4.1.0",
    "black==23.12.0",
    "ruff==0.1.8",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["anip*"]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"

